rename folder from byoc_op to byoc_i
